fix(practice): harden subject mode, db writes, and validation edge cases
- Fix pgvector query with sql.raw and add limit(20) for subject search
- Check multiline before cleanSentence; expand targetSentence tests
- Wrap page/key_events insert in transaction; widen seq/hold_duration_ms to integer
- Use UUID for generated targets and test run IDs; fix PracticePanel loading subscription
- Dedupe sessionService targetText lookup; sync TYPING_MODES and DB_SCHEMA docs

+++ src/db/schema.ts
@@ -105,12 +105,12 @@ export const keyEvents = pgTable(
     pageId: uuid("page_id")
       .references(() => pages.id, { onDelete: "cascade" })
       .notNull(),
-    seq: smallint("seq").notNull(),
+    seq: integer("seq").notNull(),
     fromKey: varchar("from_key", { length: 20 }),
     toKey: varchar("to_key", { length: 20 }).notNull(),
     keyChar: varchar("key_char", { length: 10 }).default(""),
     latency: integer("latency").notNull(),
-    holdDurationMs: smallint("hold_duration_ms"),
+    holdDurationMs: integer("hold_duration_ms"),
     isCorrect: boolean("is_correct"),
     expectedChar: varchar("expected_char", { length: 10 }),
     createdAt: timestamp("created_at", { withTimezone: true }).defaultNow().notNull(),
